--- dync/time.go
@@ -1,59 +1,0 @@
-/*
- * Copyright 2012-2024 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package dync
-
-import (
-	"encoding/json"
-	"time"
-
-	"github.com/go-spring/spring-core/conf"
-	"github.com/spf13/cast"
-	"go.uber.org/atomic"
-)
-
-type Time struct {
-	v atomic.Time
-}
-
-func (x *Time) Value() time.Time {
-	return x.v.Load()
-}
-
-func (x *Time) getTime(prop conf.ReadOnlyProperties, param conf.BindParam) (time.Time, error) {
-	s, err := GetProperty(prop, param)
-	if err != nil {
-		return time.Time{}, err
-	}
-	v, err := cast.ToTimeE(s)
-	if err != nil {
-		return time.Time{}, err
-	}
-	return v, nil
-}
-
-func (x *Time) OnRefresh(prop conf.ReadOnlyProperties, param conf.BindParam) error {
-	v, err := x.getTime(prop, param)
-	if err != nil {
-		return err
-	}
-	x.v.Store(v)
-	return nil
-}
-
-func (x *Time) MarshalJSON() ([]byte, error) {
-	return json.Marshal(x.Value())
-}
